+++ src-docs/src/views/guidelines/colors.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-
 import { Link } from 'react-router';
 
 import lightColors from '!!sass-vars-to-js-loader!../../../../src/global_styling/variables/_colors.scss';
@@ -23,6 +22,8 @@ import {
   EuiCopy,
   EuiRange,
   EuiFormRow,
+  EuiCode,
+  EuiSwitch,
 } from '../../../../src/components';
 
 const allowedColors = [
@@ -40,50 +41,40 @@ const allowedColors = [
   'euiColorAccent',
 ];
 
-const visColors = [
-  'euiColorVis0',
-  'euiColorVis1',
-  'euiColorVis2',
-  'euiColorVis3',
-  'euiColorVis4',
-  'euiColorVis5',
-  'euiColorVis6',
-  'euiColorVis7',
-  'euiColorVis8',
-  'euiColorVis9',
-];
-
 const ratingAAA = <EuiBadge color="#000">AAA</EuiBadge>;
 
 const ratingAA = <EuiBadge color="#333">AA</EuiBadge>;
 
 const ratingAA18 = <EuiBadge color="#666">AA18</EuiBadge>;
 
-function renderPaletteColor(palette, color, index) {
+function renderPaletteColor(palette, color, index, key) {
+  const hex = key ? palette[color][key] : palette[color];
+  const name = key && key !== 'graphic' ? `${color}_${key}` : color;
+
   return (
-    <EuiFlexItem key={index}>
+    <EuiFlexItem key={index} grow={false}>
       <EuiFlexGroup responsive={false} alignItems="center">
         <EuiFlexItem grow={false}>
-          <EuiCopy beforeMessage="Click to copy color name" textToCopy={color}>
+          <EuiCopy beforeMessage="Click to copy color name" textToCopy={name}>
             {copy => (
               <EuiIcon
                 onClick={copy}
                 size="xl"
                 type="stopFilled"
-                color={rgbToHex(palette[color].rgba)}
+                color={rgbToHex(hex.rgba)}
               />
             )}
           </EuiCopy>
         </EuiFlexItem>
         <EuiFlexItem>
           <EuiTitle size="xxs">
-            <h3>{color}</h3>
+            <h3>{name}</h3>
           </EuiTitle>
         </EuiFlexItem>
         <EuiFlexItem>
           <EuiText size="s" color="subdued">
             <p>
-              <code>{rgbToHex(palette[color].rgba).toUpperCase()}</code>
+              <code>{rgbToHex(hex.rgba).toUpperCase()}</code>
             </p>
           </EuiText>
         </EuiFlexItem>
@@ -91,7 +82,7 @@ function renderPaletteColor(palette, color, index) {
           <EuiText size="s" color="subdued">
             <p>
               <code>
-                rgb({palette[color].r}, {palette[color].g}, {palette[color].b})
+                rgb({hex.r}, {hex.g}, {hex.b})
               </code>
             </p>
           </EuiText>
@@ -144,6 +135,7 @@ export default class extends Component {
 
     this.state = {
       value: '3',
+      behindTextVariant: false,
     };
   }
 
@@ -153,8 +145,14 @@ export default class extends Component {
     });
   };
 
+  onBehindTextVariantChange = e => {
+    this.setState({
+      behindTextVariant: e.target.checked,
+    });
+  };
+
   render() {
-    const { value } = this.state;
+    const { value, behindTextVariant } = this.state;
     const { selectedTheme } = this.props;
 
     let palette;
@@ -166,6 +164,41 @@ export default class extends Component {
       palette = darkColors;
     } else {
       palette = lightColors;
+    }
+
+    // Vis colors are the same for all palettes
+    const visColors = lightColors.euiPaletteColorBlind;
+    const visColorKeys = Object.keys(lightColors.euiPaletteColorBlind);
+
+    function getContrastRatings(color1, color2) {
+      if (color1.indexOf('Shade') === -1 && color2.indexOf('Shade') === -1) {
+        // Exit out, i.e. don't render, if non-shade top of non-shade
+        return;
+      }
+
+      const contrast = calculateContrast(
+        [palette[color1].r, palette[color1].g, palette[color1].b],
+        [palette[color2].r, palette[color2].g, palette[color2].b]
+      );
+
+      let contrastRating;
+      let contrastRatingBadge;
+      if (contrast >= 7) {
+        contrastRating = <EuiIcon type="checkInCircleFilled" />;
+        contrastRatingBadge = ratingAAA;
+      } else if (contrast >= 4.5) {
+        contrastRating = <EuiIcon type="checkInCircleFilled" />;
+        contrastRatingBadge = ratingAA;
+      } else if (contrast >= 3) {
+        contrastRating = <EuiIcon type="editorBold" />;
+        contrastRatingBadge = ratingAA18;
+      } else if (color2.includes('Shade') && contrast >= 2) {
+        contrastRating = <EuiIcon type="minusInCircle" />;
+      } else {
+        contrastRating = <EuiIcon type="cross" />;
+      }
+
+      return { contrast, contrastRating, contrastRatingBadge };
     }
 
     return (
@@ -213,7 +246,8 @@ export default class extends Component {
             </li>
             <li>
               <EuiIcon type="editorBold" />: {ratingAA18} Passes with a contrast
-              of 3+, but only if the text is at least 18px or 14px and bold
+              of 3+, but only for graphics or if the text is at least 18px, or
+              14px and bold
             </li>
             <li>
               <EuiIcon type="minusInCircle" /> : Use only for disabled or
@@ -254,31 +288,17 @@ export default class extends Component {
                 <EuiText size="xs">
                   <h3>{color}</h3>
                   {allowedColors.map(function(color2, index) {
-                    const contrast = calculateContrast(
-                      [palette[color].r, palette[color].g, palette[color].b],
-                      [palette[color2].r, palette[color2].g, palette[color2].b]
-                    );
+                    const contrastRatings = getContrastRatings(color, color2);
 
-                    if (contrast < value) {
+                    if (!contrastRatings || contrastRatings.contrast < value) {
                       return;
                     }
 
-                    let contrastRating;
-                    let contrastRatingBadge;
-                    if (contrast >= 7) {
-                      contrastRating = <EuiIcon type="checkInCircleFilled" />;
-                      contrastRatingBadge = ratingAAA;
-                    } else if (contrast >= 4.5) {
-                      contrastRating = <EuiIcon type="checkInCircleFilled" />;
-                      contrastRatingBadge = ratingAA;
-                    } else if (contrast >= 3) {
-                      contrastRating = <EuiIcon type="editorBold" />;
-                      contrastRatingBadge = ratingAA18;
-                    } else if (color2.includes('Shade') && contrast >= 2) {
-                      contrastRating = <EuiIcon type="minusInCircle" />;
-                    } else {
-                      contrastRating = <EuiIcon type="cross" />;
-                    }
+                    const {
+                      contrast,
+                      contrastRating,
+                      contrastRatingBadge,
+                    } = getContrastRatings(color, color2);
 
                     const tooltipContent = (
                       <div>
@@ -324,20 +344,44 @@ color: $${color2};`;
           <h2>Categorical visualization palette</h2>
           <p>
             The following colors are color-blind safe and should be used in
-            categorically seried visualizations.
+            categorically seried visualizations and graphics. They are meant to
+            be contrasted against the value of{' '}
+            <EuiCode>euiColorEmptyShade</EuiCode> for the current theme.
           </p>
           <p>
             For more visualization palettes and rendering services, go to the{' '}
             <Link to="/utilities/color-palettes">Color Palettes</Link> utility
             page.
           </p>
+          <p>
+            When using the palette as a background for text (i.e. badges), use
+            the <EuiCode>_behindText</EuiCode> variant. It is a brightened
+            version of the base palette to create better contrast with text.
+          </p>
         </EuiText>
 
         <EuiSpacer />
 
+        <EuiSwitch
+          label={
+            <span>
+              Show <EuiCode>_behindText</EuiCode> variant
+            </span>
+          }
+          checked={behindTextVariant}
+          onChange={this.onBehindTextVariantChange}
+        />
+
+        <EuiSpacer />
+
         <EuiFlexGroup direction="column" gutterSize="s">
-          {visColors.map(function(color, index) {
-            return renderPaletteColor(palette, color, index);
+          {visColorKeys.map(function(color, index) {
+            return renderPaletteColor(
+              visColors,
+              color,
+              index,
+              behindTextVariant ? 'behindText' : 'graphic'
+            );
           })}
         </EuiFlexGroup>
       </GuidePage>