+++ src/components/form/checkbox/checkbox_group.tsx
@@ -1,19 +1,36 @@
-import React, { ReactNode, FunctionComponent } from 'react';
+import React, { FunctionComponent, HTMLAttributes } from 'react';
+import classNames from 'classnames';
 
-import { EuiCheckbox } from './checkbox';
-import { CommonProps } from '../../common';
+import { CommonProps, ExclusiveUnion } from '../../common';
 
-export interface EuiCheckboxGroupOption {
+import {
+  EuiFormFieldsetProps,
+  EuiFormLegendProps,
+  EuiFormFieldset,
+} from '../form_fieldset';
+import { EuiCheckbox, EuiCheckboxProps } from './checkbox';
+
+export interface EuiCheckboxGroupOption
+  extends Omit<EuiCheckboxProps, 'checked' | 'onChange'> {
   id: string;
-  label?: ReactNode;
-  disabled?: boolean;
 }
 
 export interface EuiCheckboxGroupIdToSelectedMap {
   [id: string]: boolean;
 }
 
-export interface EuiCheckboxGroupProps extends CommonProps {
+// Must omit inherit `onChange` properties or else TS complaines when applying to the EuiRadio
+type AsDivProps = Omit<HTMLAttributes<HTMLDivElement>, 'onChange'>;
+type WithLegendProps = Omit<EuiFormFieldsetProps, 'onChange'> & {
+  /**
+   * If the individual labels for each radio do not provide a sufficient description, add a legend.
+   * Wraps the group in a `EuiFormFieldset` which adds an `EuiLegend` for titling the whole group.
+   * Accepts an `EuiFormLegendProps` shape.
+   */
+  legend?: EuiFormLegendProps;
+};
+
+export type EuiCheckboxGroupProps = CommonProps & {
   options: EuiCheckboxGroupOption[];
   idToSelectedMap: EuiCheckboxGroupIdToSelectedMap;
   onChange: (optionId: string) => void;
@@ -23,7 +40,7 @@ export interface EuiCheckboxGroupProps extends CommonProps {
    */
   compressed?: boolean;
   disabled?: boolean;
-}
+} & ExclusiveUnion<AsDivProps, WithLegendProps>;
 
 export const EuiCheckboxGroup: FunctionComponent<EuiCheckboxGroupProps> = ({
   options = [],
@@ -32,22 +49,45 @@ export const EuiCheckboxGroup: FunctionComponent<EuiCheckboxGroupProps> = ({
   className,
   disabled,
   compressed,
+  legend,
   ...rest
-}) => (
-  <div className={className} {...rest}>
-    {options.map((option, index) => {
-      return (
-        <EuiCheckbox
-          className="euiCheckboxGroup__item"
-          key={index}
-          id={option.id}
-          checked={idToSelectedMap[option.id]}
-          label={option.label}
-          disabled={disabled || option.disabled}
-          onChange={onChange.bind(null, option.id)}
-          compressed={compressed}
-        />
-      );
-    })}
-  </div>
-);
+}) => {
+  const checkboxes = options.map((option, index) => {
+    const {
+      disabled: isOptionDisabled,
+      className: optionClass,
+      ...optionRest
+    } = option;
+    return (
+      <EuiCheckbox
+        className={classNames('euiCheckboxGroup__item', optionClass)}
+        key={index}
+        checked={idToSelectedMap[option.id]}
+        disabled={disabled || isOptionDisabled}
+        onChange={onChange.bind(null, option.id)}
+        compressed={compressed}
+        {...optionRest}
+      />
+    );
+  });
+
+  if (!!legend) {
+    // Be sure to pass down the compressed option to the legend
+    legend.compressed = compressed;
+
+    return (
+      <EuiFormFieldset
+        className={className}
+        legend={legend}
+        {...rest as EuiFormFieldsetProps}>
+        {checkboxes}
+      </EuiFormFieldset>
+    );
+  }
+
+  return (
+    <div className={className} {...rest as HTMLAttributes<HTMLDivElement>}>
+      {checkboxes}
+    </div>
+  );
+};