+++ src/services/color/eui_palettes.ts
@@ -1,6 +1,7 @@
+import { flatten } from 'lodash';
+import chroma from 'chroma-js';
 import { HEX } from './color_types';
 import { colorPalette } from './color_palette';
-import { flatten } from 'lodash';
 
 export type EuiPalette = string[];
 
@@ -32,15 +33,15 @@ export const euiPaletteColorBlind = function(
   let colors: string[] = [];
 
   const base = [
-    '#5BBAA0', // 0 green
+    '#54B399', // 0 green
     '#6092C0', // 1 blue
     '#D36086', // 2 dark pink
     '#9170B8', // 3 purple
-    '#EEAFCF', // 4 light pink
-    '#FAE181', // 5 yellow
-    '#CDBD9D', // 6 tan
-    '#F19F58', // 7 orange
-    '#B46F5F', // 8 brown
+    '#CA8EAE', // 4 light pink
+    '#D6BF57', // 5 yellow
+    '#B9A888', // 6 tan
+    '#DA8B45', // 7 orange
+    '#AA6556', // 8 brown
     '#E7664C', // 9 red
   ];
 
@@ -62,6 +63,20 @@ export const euiPaletteColorBlind = function(
   }
 
   return colors;
+};
+
+/**
+ * Color blind palette with text is meant for use when text is applied on top of the color.
+ * It increases the brightness of the color to give the text more contrast.
+ */
+export const euiPaletteColorBlindBehindText = function(): EuiPalette {
+  const originalPalette = euiPaletteColorBlind();
+  const newPalette = originalPalette.map(color =>
+    chroma(color)
+      .brighten(0.5)
+      .hex()
+  );
+  return newPalette;
 };
 
 export const euiPaletteForLightBackground = function(): EuiPalette {